refacto (feature) : split GeoJSon2THREE
split GeoJSon2THREE in 3 parts

* GeoJSON2Feature transforms geojson to itowns structure
* Feature2Mesh transforms itowns structure to THREE.Mesh
* Feature2Texture transforms itowns structure to THREE.Texture
