+++ src/Core/Geographic/Coordinates.js
@@ -423,19 +423,7 @@ Coordinates.prototype.z = function z() {
 
 Coordinates.prototype.xyz = function xyz(target) {
     _assertIsGeocentric(this.crs);
-    return this.rawValues(target);
-};
-
-/**
- * The function returns the coordinates's values without convertion in Vector3
- * Vector3.x: longitude or x value
- * Vector3.y: latitude or y value
- * Vector3.z: altitude or z value
- * @param      {Vector3}  optionalTarget  if there's optionalTarget the result is copy in it
- * @return     {Vector3}  return the coordinates's values without convertion in Vector3
- */
-Coordinates.prototype.rawValues = function rawValues(optionalTarget) {
-    const v = optionalTarget || new THREE.Vector3();
+    const v = target || new THREE.Vector3();
     v.fromArray(this._values);
     return v;
 };
